feat(wasm): start to implement keystore for browser
