+++ internal/pkg/crypto/keystore_native.go
@@ -1,0 +1,11 @@
+//go:build !js
+// +build !js
+
+package crypto
+
+func InitKeystore(KeyStoreName, KeyStoreDir string) (int, error) {
+	signkeycount := 0
+	var err error
+	ks, signkeycount, err = InitDirKeyStore(KeyStoreName, KeyStoreDir)
+	return signkeycount, err
+}
